Added logging, excluding game directory

+++ server/server.py
@@ -3,19 +3,23 @@ from _stats import Stats
 import json
 import subprocess
 import os
+from datetime import datetime # why is there datetime inside datetime??
 from gamemap import *
 
 class Game:
-    def __init__(self, map_:Map):
+    def __init__(self, map_:Map, log):
         self.map = map_
+        self.log = log
     def executeTurns(self, turns):
         for playerID, turn in enumerate(turns):
             if self.validateTurn(playerID, turn):
                 self.executeTurn(playerID, turn)
             else:
-                print(f"Invalid turn from player: {playerID}")
+                self.log(f"Invalid turn from player: {playerID}")
     def parse(self):
         return self.map.parse()
+    def removePlayer(self, playerID):
+        pass #TODO
 
 with open(os.path.join(os.path.dirname(__file__), '..', 'config.json')) as f:
     CONFIG_JSON = json.load(f)
@@ -36,47 +40,70 @@ def setupGameDir(CONFIG_JSON:dict, GAMES_JSON:dict):
             if file[:-len('.log')] == '.log':
                 os.remove(os.path.join(logdir, file))
     return gamedir, logdir
+LOG_FILE = None
+def setupLogger(logdir):
+    global LOG_FILE
+    LOG_FILE = open(os.path.join(logdir, '__server__.log'), 'w')
+    def log(*msg):
+        message = ' '.join([str(i) for i in msg]) + '\n'
+        LOG_FILE.write(message)
+        print(message, end='')
+    return log
 
 gamedir, logdir = setupGameDir(CONFIG_JSON, GAMES_JSON)
 
-player_programs = [(key, value['path'], value['command']) for key, value in CONFIG_JSON.items()]
+log = setupLogger(logdir)
+log(f'[{datetime.now().strftime('%d.%B.%Y-%H:%M:%S')}] Start of log file')
+
+player_program_files = [(key, value['path'], value['command']) for key, value in CONFIG_JSON.items()]
 
 player_subprocesses = {
     # subprocess.Popen(['python3', 'bot.py'], stdin=subprocess.PIPE, stdout=subprocess.PIPE, text=True),
     # subprocess.Popen(['python3', 'bot.py'], stdin=subprocess.PIPE, stdout=subprocess.PIPE, text=True)
 }
 
-for id, path, command in player_programs:
+for id, path, command in player_program_files:
     with open(os.path.join(logdir, str(id) + '.log'), 'w') as stderr_output:
         player_subprocesses[id] = subprocess.Popen([command, path], stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=stderr_output, text=True)
 
 def handle_timeout_or_error(p):
-    print(f"Timeout: {p}")
+    log(f"Timeout: {p}")
 
 m = Map()
 m.new()
-g = Game(m)
+g = Game(m, log)
 game_active = True
 turn = 0
+players_errored_out:dict[int, Exception] = {}
 while game_active:
     state = g.parse()
     moves = []
     for player_id in player_subprocesses:
+        if player_id in players_errored_out:
+            continue
         player = player_subprocesses[player_id]
         # Send data
-        player.stdin.write(state + "\n")
-        player.stdin.flush()
+        try:
+            player.stdin.write(state + "\n")
+            player.stdin.flush()
+        except BrokenPipeError:
+            log(f"Error getting response from player {player_id}: Broken pipe")
+            log(f"Removing player from game...")
+            g.removePlayer(player_id)
+            player_subprocesses[player_id].terminate()
+            players_errored_out[player_id] = BrokenPipeError
         # Get response with timeout logic
         try:
             move = player.stdout.readline().strip()
             moves.append((move, player))
         except Exception as e:
-            print(f"Error {e} occured while playing turn of player {player}")
+            log(f"Error {e} occured while playing turn of player {player}")
             handle_timeout_or_error(player)
-        print(f'Heppy turns! {moves}')
+        log(f'Heppy turns! {moves}')
     turn += 1
     if turn > 20:
-        print("\n\nFOR TESTING PURPOSES TURNS WERE CAPPED TO 20\nEnding game...\n\n")
+        log("\n\nFOR TESTING PURPOSES TURNS WERE CAPPED TO 20\nEnding game...\n\n")
         break
-for player_subprocess in player_subprocesses.values():
-    player_subprocess.terminate()
+for key, player_subprocess in player_subprocesses.items():
+    if not key in players_errored_out:
+        player_subprocess.terminate()
